move tooltip context to app context
