+++ packages/dashboard/src/components/app-contexts.tsx
@@ -8,18 +8,26 @@ export const ResourcesContext = React.createContext<ResourceManager | undefined>
 
 export const SettingsContext = React.createContext(defaultSettings());
 
+export const TooltipContext = React.createContext({
+  showTooltips: false,
+  toggleTooltips: () => {},
+});
+
 export interface AppContextProviderProps extends React.PropsWithChildren<{}> {
   settings: Settings;
   resourceManager?: ResourceManager;
   notificationDispatch: React.Dispatch<React.SetStateAction<NotificationBarProps | null>>;
+  tooltips: { showTooltips: boolean; toggleTooltips: () => void };
 }
 
 export function AppContextProvider(props: AppContextProviderProps): React.ReactElement {
-  const { settings, notificationDispatch, resourceManager, children } = props;
+  const { settings, notificationDispatch, resourceManager, tooltips, children } = props;
   return (
     <SettingsContext.Provider value={settings}>
       <NotificationBarContext.Provider value={notificationDispatch}>
-        <ResourcesContext.Provider value={resourceManager}>{children}</ResourcesContext.Provider>
+        <TooltipContext.Provider value={tooltips}>
+          <ResourcesContext.Provider value={resourceManager}>{children}</ResourcesContext.Provider>
+        </TooltipContext.Provider>
       </NotificationBarContext.Provider>
     </SettingsContext.Provider>
   );
