--- packages/dashboard/src/components/ui-contexts.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-
-export const TooltipContext = React.createContext({
-  showTooltips: true,
-  toggleTooltips: () => {},
-});
